+++ src/config/AxiosHelper.js
@@ -1,5 +1,9 @@
 import axios from "axios";
 export const baseURL = "http://localhost:8080";
 export const httpClient = axios.create({
-  baseURL: baseURL
+    baseURL: baseURL,
+    headers:{
+        'Access-Control-Allow-Origin': '*',
+
+    }
 });
