+++ src/config/helper.js
@@ -1,31 +1,20 @@
+const helper =  function helper(date) {
+    const now = new Date();
+    const past = new Date(date);
+    const secondsAgo = Math.floor((now - past) / 1000);
 
-export function timeAgo(date){
-    const currentDate = new Date();
-    const givenDate = new Date(date);
-    const difference = currentDate - givenDate;
-    const seconds = difference / 1000;
-    if (seconds < 60) {
-        return `${Math.floor(seconds)} seconds ago`;
-    }
-    const minutes = seconds / 60;
-    if (minutes < 60) {
-        return `${Math.floor(minutes)} minutes ago`;
-    }
-    const hours = minutes / 60;
-    if (hours < 24) {
-        return `${Math.floor(hours)} hours ago`;
-    }
-    const days = hours / 24;
-    if (days < 30) {
-        return `${Math.floor(days)} days ago`;
-    }
-    const months = days / 30;
-    if (months < 12) {
-        return `${Math.floor(months)} months ago`;
-    }
-    const years = months / 12;
-    return `${Math.floor(years)} years ago`;
+    if (secondsAgo < 60) return `${secondsAgo} seconds ago`;
+    const minutesAgo = Math.floor(secondsAgo / 60);
+    if (minutesAgo < 60) return `${minutesAgo} minutes ago`;
+    const hoursAgo = Math.floor(minutesAgo / 60);
+    if (hoursAgo < 24) return `${hoursAgo} hours ago`;
+    const daysAgo = Math.floor(hoursAgo / 24);
+    if (daysAgo < 30) return `${daysAgo} days ago`;
+    const monthsAgo = Math.floor(daysAgo / 30);
+    if (monthsAgo < 12) return `${monthsAgo} months ago`;
+    const yearsAgo = Math.floor(monthsAgo / 12);
+    return `${yearsAgo} years ago`;
 }
-
+export default helper;
 // Example usage
-console.log(timeAgo("2023-12-01T14:00:00Z")); // Output depends on the current time
+console.log(helper("2023-12-01T14:00:00Z"));
